+++ R/algo_tree_detection.R
@@ -39,10 +39,8 @@
 #' If it is a function, the function determines the size of the window at any given location on the canopy.
 #' The function should take the height of a given pixel or points as its only argument and return the
 #' desired size of the search window when centered on that pixel/point.
-#'
 #' @param hmin numeric. Minimum height of a tree. Threshold below which a pixel or a point
 #' cannot be a local maxima. Default is 2.
-#'
 #' @param shape character. Shape of the moving windows used to find the local maxima. Can be "square"
 #' or "circular".
 #'
@@ -57,43 +55,43 @@
 #'
 #' @examples
 #' LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
-#' las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
+#' las <- readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
 #'
 #' # point-cloud-based
 #' # =================
 #'
 #' # 5x5 m fixed windows size
-#' ttops = tree_detection(las, lmf(5))
+#' ttops <- tree_detection(las, lmf(5))
 #'
-#' x = plot(las)
+#' x <- plot(las)
 #' add_treetops3d(x, ttops)
 #'
 #' # variable windows size
-#' f = function(x) { x * 0.07 + 3}
-#' ttops = tree_detection(las, lmf(f))
+#' f <- function(x) { x * 0.07 + 3}
+#' ttops <- tree_detection(las, lmf(f))
 #'
-#' x = plot(las)
+#' x <- plot(las)
 #' add_treetops3d(x, ttops)
 #'
 #' # raster-based
 #' # ============
 #'
 #' # 5x5 m fixed windows size
-#' chm = grid_canopy(las, res = 1, p2r(0.15))
-#' kernel = matrix(1,3,3)
-#' chm = raster::focal(chm, w = kernel, fun = median, na.rm = TRUE)
+#' chm <- grid_canopy(las, res = 1, p2r(0.15))
+#' kernel <- matrix(1,3,3)
+#' chm <- raster::focal(chm, w = kernel, fun = median, na.rm = TRUE)
 #'
-#' ttops = tree_detection(chm, lmf(5))
+#' ttops <- tree_detection(chm, lmf(5))
 #'
-#' raster::plot(chm, col = height.colors(30))
-#' sp::plot(ttops, add = TRUE)
+#' plot(chm, col = height.colors(30))
+#' plot(ttops, add = TRUE)
 #'
 #' # variable windows size
-#' f = function(x) { x * 0.07 + 3 }
-#' ttops = tree_detection(chm, lmf(f))
+#' f <- function(x) { x * 0.07 + 3 }
+#' ttops <- tree_detection(chm, lmf(f))
 #'
-#' raster::plot(chm, col = height.colors(30))
-#' sp::plot(ttops, add = TRUE)
+#' plot(chm, col = height.colors(30))
+#' plot(ttops, add = TRUE)
 lmf = function(ws, hmin = 2, shape = c("circular", "square"))
 {
   shape <- match.arg(shape)
@@ -106,37 +104,27 @@ lmf = function(ws, hmin = 2, shape = c("circular", "square"))
     context <- tryCatch({get("lidR.context", envir = parent.frame())}, error = function(e) {return(NULL)})
     stopif_wrong_context(context, "tree_detection", "lmf")
 
-    n = nrow(las@data)
-
-    if (is.numeric(ws))
+    if (is.function(ws))
     {
-      # nothing to do
-    }
-    else if (is.function(ws))
-    {
-      ws = ws(las@data$Z)
-      ws[las@data$Z < hmin] = ws(hmin)
+      n     <- nrow(las@data)
+      ws    <- ws(las@data$Z)
+      b     <- las$Z < hmin
+      ws[b] <- ws(hmin)
 
       if (!is.numeric(ws)) stop("The function 'ws' did not return a correct output. ", call. = FALSE)
       if (any(ws <= 0))    stop("The function 'ws' returned negative or null values.", call. = FALSE)
-      if (anyNA(ws))       stop("The function 'ws' returned NA values.", call. = FALSE)
-      if (length(ws) != n) stop("The function 'ws' did not return a correct output.", call. = FALSE)
+      if (anyNA(ws))       stop("The function 'ws' returned NA values.",               call. = FALSE)
+      if (length(ws) != n) stop("The function 'ws' did not return a correct output.",  call. = FALSE)
     }
-    else
+    else if (!is.numeric(ws))
+    {
       stop("'ws' must be a number or a function", call. = FALSE)
+    }
 
-    . <- X <- Y <- Z <- treeID <- NULL
-    is_maxima = C_lmf(las@data, ws, hmin, circ, getThread())
-    maxima = las@data[is_maxima, .(X,Y,Z)]
-    maxima[, treeID := 1:.N]
-
-    output = sp::SpatialPointsDataFrame(maxima[, .(X,Y)], maxima[, .(treeID, Z)])
-    output@proj4string = las@proj4string
-    output@bbox = sp::bbox(las)
-    return(output)
+    return(C_lmf(las@data, ws, hmin, circ, getThread()))
   }
 
-  class(f) <- c("PointCloudBased", "IndividualTreeDetection", "Algorithm", "lidR")
+  class(f) <- c("PointCloudBased", "IndividualTreeDetection", "OpenMP", "Algorithm", "lidR")
   return(f)
 }