+++ R/catalog_fakerun.r
@@ -1,4 +1,4 @@
-catalog_fakerun = function(ctg, sleep = 0.05)
+catalog_fakerun = function(ctg, sleep = 0.05, pwarning = 0.05, perror = 0.01)
 {
   opt_wall_to_wall(ctg) <- FALSE
   options <- list(need_buffer = FALSE, drop_null = FALSE)
@@ -6,11 +6,10 @@ catalog_fakerun = function(ctg, sleep = 0.05)
   fake = function(x)
   {
     Sys.sleep(sleep)
-    if (stats::runif(1) < 0.05) warning("Fake warning", call. = FALSE)
-    if (stats::runif(1) < 0.01) stop("Fake error", call. = FALSE)
+    if (stats::runif(1) < pwarning) warning("Fake warning", call. = FALSE)
+    if (stats::runif(1) < perror) stop("Fake error", call. = FALSE)
     return(0)
   }
 
-  catalog_apply(ctg, fake, .options = options)
-  return(invisible())
+  return(catalog_apply(ctg, fake, .options = options))
 }