+++ R/grid_metrics.r
@@ -211,7 +211,8 @@ grid_metrics.LAScatalog = function(las, func, res = 20, start = c(0,0), filter =
   is_formula <- tryCatch(lazyeval::is_formula(func), error = function(e) FALSE)
   if (!is_formula) func <- lazyeval::f_capture(func)
 
-  options <- list(need_buffer = FALSE, drop_null = TRUE, raster_alignment = alignment)
+  globals <- future::getGlobalsAndPackages(func)
+  options <- list(need_buffer = FALSE, drop_null = TRUE, globals = names(globals$globals), raster_alignment = alignment)
   output  <- catalog_apply(las, grid_metrics, func = func, res = res, start = start, filter = filter, .options = options)
 
   if (opt_output_files(las) != "")                # Outputs have been written in files. Return a virtual raster mosaic