+++ R/utils_is.r
@@ -25,12 +25,20 @@
 #
 # ===============================================================================
 
-#' Test if a \code{LAS} object is empty
+#' A set of boolean tests on objects
 #'
-#' An empty \code{LAS} object is a point cloud with 0 points
+#' \code{is.empty} tests if a \code{LAS} object is a point cloud with 0 points.\cr
+#' \code{is.overlapping} tests if a \code{LAScatalog} have overlapping tiles.\cr
+#' \code{is.indexed} tests if the points of a \code{LAScatalog} are indexed with \code{.lax} files.\cr
+#' \code{is.algorithm} tests if an object is an algorithm of the lidR package.\cr
+#' \code{is.parallelised} tests if an algorithm of the lidR package is natively parallelised with OpenMP.
+#' Returns TRUE if the algorithm is at least partially parallelised i.e. if some portion of the code are
+#' computed in parallel.
 #'
-#' @param object A \code{LAS} object
-#' @param ... Unused
+#' @param las A \code{LAS} object
+#' @param x Any R object.
+#' @param catalog A \code{LAScatalog} object
+#' @param algorithm An \code{algorithm} object
 #'
 #' @return TRUE or FALSE
 #'
@@ -41,19 +49,30 @@
 #'
 #' las = new("LAS")
 #' is.empty(las)
+#'
+#' f <- lmf(2)
+#' is.parallelised(f)
+#'
+#' g <- pitfree()
+#' is.parallelised(g)
+#'
+#' ctg <- catalog(LASfile)
+#' is.indexed(ctg)
 #' @export
-setGeneric("is.empty", function(object, ...)
-  standardGeneric("is.empty"))
+#' @rdname is
+#' @name is
+NULL
 
+#' @rdname is
 #' @export
-#' @rdname is.empty
-setMethod("is.empty", "LAS", function(object, ...)
+is.empty <- function(las)
 {
-  empty = if (nrow(object@data) == 0) TRUE else FALSE
-  return(empty)
-})
+  stopifnotlas(las)
+  return(nrow(las@data) == 0L)
+}
 
-
+#' @rdname is
+#' @export
 is.overlapping = function(catalog)
 {
   contour       <- rgeos::gUnaryUnion(catalog)
@@ -62,9 +81,27 @@ is.overlapping = function(catalog)
   return(actual_area < measured_area)
 }
 
-
+#' @rdname is
+#' @export
 is.indexed = function(catalog)
 {
   laxfiles <- paste0(tools::file_path_sans_ext(catalog@data$filename), ".lax")
   return(!any(!file.exists(laxfiles)))
+}
+
+#' @rdname is
+#' @export
+is.algorithm = function(x)
+{
+  return(is(x, "lidR") && is(x, "Algorithm"))
+}
+
+#' @rdname is
+#' @export
+is.parallelised = function(algorithm)
+{
+  if (!is.algorithm(algorithm))
+    stop("This function only applies to algorithms from lidR package")
+
+  return(is(algorithm, "OpenMP"))
 }